Debugged Dockerfile - Debbuged Dockerfile (still not working) - Debugged gunicorn app (still not working)
