+++ Dockerfile
@@ -6,8 +6,10 @@
 
 # Want to help us make this template better? Share your feedback here: https://forms.gle/ybq9Krt8jtBL3iCk7
 
-ARG PYTHON_VERSION=3.10.12
-FROM python:${PYTHON_VERSION}-slim as base
+# ARG PYTHON_VERSION=3.10.12
+# FROM python:${PYTHON_VERSION}-slim as base
+FROM python:latest as base
+
 
 # Prevents Python from writing pyc files.
 ENV PYTHONDONTWRITEBYTECODE=1
@@ -48,4 +50,5 @@ COPY . .
 EXPOSE 8000
 
 # Run the application.
-CMD ["gunicorn", "-c", "gunicorn_config.py", "falafel.wsgi:create_app()"]
+CMD ["gunicorn", "--config", "falafel/gunicorn_conf.py"]
+#CMD ["gunicorn", "-c", "falafel/gunicorn_conf.py", "falafel.wsgi:create_app()"]
