+++ compose.yaml
@@ -29,9 +29,9 @@ services:
   nginx:
       image: nginx:latest
       ports:
-        - "8000:8000"
+        - "5000:5000"
       volumes:
-        - ./nginx.conf:/etc/nginx/conf.d/default.conf
+        - ./config/nginx.conf:/etc/nginx/conf.d/default.conf
       depends_on:
         - falafel
       restart: unless-stopped
