+++ config/falafel.conf
@@ -3,7 +3,7 @@
 # OS : Case Sensetive, only 3 accepted values : 'Linux' or 'Windows' or 'WSL'
 [System]
 
-OS = WSL
+OS = Linux
 
 
 #### Parameters used to create and train DL model
