+++ config/nginx.conf
@@ -1,0 +1,12 @@
+server {
+    listen 5000;                    # Port on which NGINX server listens (from external)
+    server_name falafel.nawfal.ca;  # Change to your domain or use `_` for any
+
+    location / {
+        proxy_pass http://falafel:8000;                 # forwards request to backend server which gunicorn server
+        proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;        # Allows beckend server to know the actual IP of the client
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;    # This header can be used to track the original IP address of a client when the request passes through multiple proxies
+        proxy_set_header X-Forwarded-Proto $scheme;     # This is useful for identifying whether the original request was made over HTTP or HTTPS
+    }
+}
