--- falafel/dl_model/models/model_CatDog.keras
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:85fd8a0b168942379b2ce7814e4126fb825f05ec5671f7e0d202f5305a2b4ba6
-size 15635524
